+++ src/brown_noise.cpp
@@ -14,7 +14,7 @@ BrownNoise::BrownNoise()
     return;
 }
 
-BrownNoise::BrownNoise(int width, int height, double leaky_integrator, int output_levels)
+BrownNoise::BrownNoise(size_t width, size_t height, double leaky_integrator, size_t output_levels)
 {
     this->threshold_matrix = std::vector<std::vector<int>>(height, std::vector<int>(width, 0));
     this->width = width;
@@ -43,15 +43,15 @@ void BrownNoise::generate_brown_noise()
     std::vector<std::vector<int>> threshold_matrix_white_noise(white_noise.get_threshold_matrix());
 
     // integrate white noise threshold matrix
-    for(int y = 0; y < height; y++)
+    for(size_t y = 0; y < height; y++)
     {
-        for(int x = 0; x < width; x++)
+        for(size_t x = 0; x < width; x++)
         {
             value = 0;
 
-            for(int j = 0; j < y; j++)
+            for(size_t j = 0; j < y; j++)
             {
-                for(int i = 0; i < x; i++)
+                for(size_t i = 0; i < x; i++)
                 {
                     value += threshold_matrix_white_noise[j][i] * leaky_integrator;
                 }
@@ -74,9 +74,9 @@ void BrownNoise::generate_brown_noise()
     value_range = value_max - value_min;
 
     // normalize to output levels
-    for(int y = 0; y < height; y++)
+    for(size_t y = 0; y < height; y++)
     {
-        for(int x = 0; x < width; x++)
+        for(size_t x = 0; x < width; x++)
         {
             threshold_matrix[y][x] = (threshold_matrix[y][x] - value_min) * (output_levels - 1) / value_range;
         }
@@ -85,57 +85,13 @@ void BrownNoise::generate_brown_noise()
     return;
 }
 
-// // generates a threshold matrix with brownian noise
-// void BrownNoise::generate_brown_noise()
-// {
-//     std::random_device rd;
-//     std::mt19937 mt(rd());
-//     std::uniform_real_distribution<> dis(-1.0, 1.0);
-//     std::vector<std::vector<double>> threshold_matrix_static_cast<double>(height, std::vector<double>(width, 0.0));
-//     double value_previous = 0.0;
-//     double value_range = 0.0;
-
-//     // populate temporary double matrix
-//     for(int y = 0; y < height; y++)
-//     {
-//         for(int x = 0; x < width; x++)
-//         {
-//             threshold_matrix_double[y][x] = leaky_integrator * (value_previous + dis(mt));
-
-//             if(threshold_matrix_double[y][x] < value_min)
-//             {
-//                 value_min = threshold_matrix_double[y][x];
-//             }
-
-//             if(threshold_matrix_double[y][x] > value_max)
-//             {
-//                 value_max = threshold_matrix_double[y][x];
-//             }
-//         }
-//     }
-
-//     value_range = value_max - value_min;
-//     std::cout << value_min << " " << value_max << " " << value_range << std::endl;
-
-//     // normalize double matrix to output levels in int matrix
-//     for(int y = 0; y < height; y++)
-//     {
-//         for(int x = 0; x < width; x++)
-//         {
-//             threshold_matrix[y][x] = static_cast<int>((threshold_matrix_double[y][x] - value_min) * static_cast<double>(output_levels) / value_range);
-//         }
-//     }
-
-//     return;
-// }
-
 std::string BrownNoise::to_string()
 {
     std::string output = "";
 
-    for(int y = 0; y < height; y++)
+    for(size_t y = 0; y < height; y++)
     {
-        for(int x = 0; x < width; x++)
+        for(size_t x = 0; x < width; x++)
         {
             output += std::to_string(threshold_matrix[y][x]) + " ";
         }
